requirements.txt
asgiref==3.5.2
Django==4.1.3
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.6
djangorestframework==3.13.1
uuid==1.30
PyJWT==2.6.0
pydantic~=1.10.2
PyYAML~=6.0
psycopg2-binary==2.9.5
django-environ==0.3.1
